Data.Group

Documentation

class Monoid m => Group m where

A Group is a Monoid plus a function, invert, such that:

a <> invert a == mempty

invert a <> a == mempty

Minimal complete definition

invert

Methods

invert :: m -> m

(~~) :: m -> m -> m infixl 7

Group subtraction: x ~~ y == x <> invert y

pow :: Integral x => m -> x -> m

pow a n == a <> a <> ... <> a

 (n lots of a)

If n is negative, the result is inverted.

class Group g => Abelian g

An Abelian group is a Group that follows the rule:

a <> b == b <> a

class Group a => Cyclic a where

A Group G is Cyclic if there exists an element x of G such that for all y in G, there exists an n, such that

y = pow x n

Methods

generator :: a

The generator of the Cyclic group.

generated :: Cyclic a => [a]

Generate all elements of a Cyclic group using its generator.

Note: Fuses, does not terminate even for finite groups.

generated' :: (Eq a, Cyclic a) => [a]

Lazily generate all elements of a Cyclic group using its generator.

Note: Fuses, terminates if the underlying group is finite.