Data.Group

Documentation

class Monoid m => Group m where

A Group is a Monoid plus a function, invert, such that:

a <> invert a == mempty

invert a <> a == mempty

Minimal complete definition

invert

Methods

invert :: m -> m

pow :: Integral x => m -> x -> m

pow a n == a <> a <> ... <> a

 (n lots of a)

If n is negative, the result is inverted.

class Group g => Abelian g

An Abelian group is a Group that follows the rule:

a <> b == b <> a